[TWINFACES-324] refactor: update imports, rename folders and refactor table fields render
