+++ src/entities/pipeline-step/libs/helpers.ts
