+++ src/entities/pipeline-step/libs/hooks/use-search-v1.ts
@@ -4,7 +4,6 @@ import { PaginationState } from "@tanstack/table-core";
 import { FactoryFilters } from "../../../factory/api/types";
 import { hydratePipelineStepFromMap } from "../helpers";
 import { PipelineStep_DETAILED } from "../../api";
-import { log } from "console";
 
 export function usePipelineStepSearch() {
   const api = useContext(ApiContext);
@@ -18,7 +17,7 @@ export function usePipelineStepSearch() {
       filters?: FactoryFilters;
     }): Promise<PagedResponse<PipelineStep_DETAILED>> => {
       try {
-        const { data, error } = await api.pipelineSteps.search({
+        const { data, error } = await api.pipelineStep.search({
           pagination,
           filters,
         });
@@ -26,8 +25,8 @@ export function usePipelineStepSearch() {
         if (error) {
           throw error;
         }
-        const pipelineSteps = (data.steps || []).map(
-          (dto) => hydratePipelineStepFromMap(dto, data.relatedObjects) ?? []
+        const pipelineSteps = (data.steps || []).map((dto) =>
+          hydratePipelineStepFromMap(dto, data.relatedObjects)
         );
 
         return {
