--- src/entities/pipelineSteps/libs/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./useSearchV1";
